pyproject.toml
[project]
name = "autoapply"
version = "1.9.0"
description = "Smart job application automation bot with Electron dashboard"
requires-python = ">=3.11"
dependencies = [
    "flask==3.1.3",
    "flask-socketio==5.6.1",
    "playwright==1.58.0",
    "pydantic==2.12.5",
    "reportlab==4.4.10",
    "requests==2.32.5",
    "keyring==25.7.0",
    "apscheduler==3.11.2",
    "httpx==0.28.1",
    "python-engineio==4.13.1",
    "python-socketio==5.16.1",
    "gevent==25.9.1",
    "gevent-websocket==0.10.1",
    "jinja2==3.1.6",
    "PyPDF2==3.0.1",
    "python-docx==1.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "ruff>=0.11",
    "pre-commit>=4.0",
    "pip-audit>=2.7",
    "mypy>=1.10",
    "types-requests>=2.31",
]

# ─── Setuptools ──────────────────────────────────────────────────────────

[tool.setuptools.packages.find]
include = ["bot*", "config*", "core*", "db*", "routes*"]

# ─── Ruff ────────────────────────────────────────────────────────────────

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    "venv",
    "electron/node_modules",
    ".pytest_cache",
    "__pycache__",
    ".git",
]

[tool.ruff.lint]
select = [
    "F",     # pyflakes
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "I",     # isort
]
ignore = [
    "E501",  # line too long — handled by formatter, not linter
]

[tool.ruff.lint.isort]
known-first-party = ["app", "app_state", "bot", "config", "core", "db", "routes"]

# ─── Pytest ──────────────────────────────────────────────────────────────

# ─── mypy ─────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "venv/",
    "electron/",
    "tests/",
]

# ─── Pytest ──────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
